requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
gspread
oauth2client
networkx
pyarrow
dune-client
fastparquet

